MSEARCH-350 Reorganize stream resource IDs IT
